RateTables

Documentation for RateTables. This package provides rate tables for person-year computations, alike R's ratetable class.

You can install it through :

using Pkg
Pkg.add("https://github.com/JuliaSurv/RateTables.jl")

Then loading it through

using RateTables

provides a constant dictionary RT_HMD from which countries rate tables can be queried as follows:

julia> RT_HMD[:SVN]
RateTable for Slovenia (code SVN):
    Ages, in years from 0 to 110 (in days from 0.0 to 40176.509999999995)
    Date, in years from 1983 to 2019 (in days from 724272.9029999999 to 737421.579)
    Sex is binary coded {0 => female, 1 => male}
You can query daily hazard values through the `daily_hazard` function.

julia>

To obtain, for example, the daily hazard rate for a male (code 1), from slovenia, on it's 80th birthday on the 1rst january 2019:

julia>  daily_hazard(RT_HMD[:SVN], 365.241*80, 365.241*2019, 1)
0.00016397140489082858

julia> 

RateTables.RT_HMD

RateTables.RateTable

RateTables.daily_hazard

RateTables.RT_HMD — Constant

`RT_HMD`

Dictionary that contains the rates tables for each country. The country codes are used to differentiate them.

These rates tables contain daily hazard rates for both sexes. They are derived from annual death probabilities (qₓ's) from the Human Mortality Database at https://www.mortality.org/

soloviart = RTHMD[:SVN]

Dayly hazard for a slovenian male (code 0), born on the 1rst of january 1927, 192 days after his 37th birthday.

daily_hazard(slovenia_rt, 37*RT_DAYS_IN_YEAR + 192, 1927*RT_DAYS_IN_YEAR + 192, 1)

RateTables.RateTable — Type

RateTable

This class deals with daily rate tables used in person-years computation. For the moment it simply holds data from the HMD.

RateTables.daily_hazard — Method

`daily_hazard(rt,age,date,sex)`

This function queries daily hazard values from a given RateTable. The parameters age and date have to be in days (1 year = 365.241 days). Parameter sex is coded 0 for females and 1 for males.