RateTables

Documentation for RateTables. This package provides rate tables for person-year computations, alike R's ratetable class.

You can install it through :

using Pkg
Pkg.add("https://github.com/JuliaSurv/RateTables.jl")

Then loading it through

using RateTables

provides a constant dictionary RT_HMD from which countries rate tables can be queried as follows:

RT_HMD[:SVN]
julia>

The list of availiable countries is as follows:

Dict(k => RT_HMD[k].country_name for k in keys(RT_HMD))

Dict{Symbol, String} with 50 entries:
  :ESP     => "Spain"
  :FIN     => "Finland"
  :GBR_NP  => "United Kingdom"
  :BEL     => "Belgium"
  :CHE     => "Switzerland"
  :DEUTE   => "East Germany"
  :LUX     => "Luxembourg"
  :IRL     => "Ireland"
  :DNK     => "Denmark"
  :GBR_NIR => "Northern Ireland"
  :CZE     => "Czechia"
  :CHL     => "Chile"
  :EST     => "Estonia"
  :ISL     => "Iceland"
  :ITA     => "Italy "
  :JPN     => "Japan"
  :NZL_NM  => "New Zealand -- Non-Maori"
  :SVK     => "Slovakia "
  :DEUTW   => "West Germany"
  ⋮        => ⋮

To obtain, for example, the daily hazard rate for a male (code 1), from slovenia, on it's 80th birthday on the 1rst january 2019:

daily_hazard(RT_HMD[:SVN], 365.241*80, 365.241*2019, 1)

0.00016397140489082858

RateTables.RT_HMD

RateTables.RateTable

RateTables.daily_hazard

RateTables.RT_HMD — Constant

`RT_HMD`

Dictionary that contains the rates tables for each country. The country codes are used to differentiate them.

These rates tables contain daily hazard rates for both sexes. They are derived from annual death probabilities (qₓ's) from the Human Mortality Database at https://www.mortality.org/

Dayly hazard for a slovenian male (code 0), born on the 1rst of january 1927, 192 days after his 37th birthday can for example be obtained as follows:

solovia_rt = RT_HMD[:SVN] daily_hazard(slovenia_rt, 37*RT_DAYS_IN_YEAR + 192, 1927*RT_DAYS_IN_YEAR + 192, 1)

The list of availiables countries can be queried via:

Dict(k => RT_HMD[k].country_name for k in keys(RT_HMD))

RateTables.RateTable — Type

RateTable

This class deals with daily rate tables used in person-years computation. For the moment it simply holds data from the HMD.

RateTables.daily_hazard — Method

`daily_hazard(rt,age,date,sex)`

This function queries daily hazard values from a given RateTable. The parameters age and date have to be in days (1 year = 365.241 days). Parameter sex is coded 0 for females and 1 for males.