RateTables

Documentation for RateTables. This package provides rate tables for person-year computations, alike R's ratetable class.

You can install it through :

using Pkg
Pkg.add("https://github.com/JuliaSurv/RateTables.jl")

Then loading this package provides a few constants rate tables objects. See the bottom of this page for the available rate tables. For the sake of the example, we'll use the hmd_rates dictionary, which stores rates tables extracted from the Human Mortality Database (HMD).

using RateTables
hmd_rates

RT(:country, :sex) with elements BRT(0..110 × 1990..2020).

To obtain, for example, the daily hazard rate for a male, slovene, on its 20th birthday on the first of january 2010, you can call the daily_hazard function. It needs to be called with the arguments in a specific format:

The age needs to be given in days, and the converting factor is 1 year = 365.241 days.

The date also needs to be given in days, same converting factor.

Other covariates formats depend on the rate table. Usually, there is a sex covariates with values :male and :female, sometimes :total too, but other covariates might have different format depending on the rate table. In particular, the country is a covariate in this particular ratetable.

For hmd_rates, there are two additional covariates: country and sex, in this order. Depending on the querying syntax, the order of the passed argument might matter:

c = :svn # slovenia.
a = 20*365.241
d = 2010*365.241
s = :male
v1 = daily_hazard(hmd_rates, a, d, c, s)
v2 = daily_hazard(hmd_rates, a, d; country=c, sex=s)
v3 = daily_hazard(hmd_rates, a, d; sex=s, country=c) # when using kwargs syntax, order of additional covariates does not matter.
v4 = daily_hazard(hmd_rates[c, s], a, d) # here the order of arguments (c,s) also matter.
(v1,v2,v3, v4)

(2.092569007129928e-6, 2.092569007129928e-6, 2.092569007129928e-6, 2.092569007129928e-6)

Note that there is a discrepency between HMD datasets and datasets from other places. For completeness, this package also include datasets commonly used in R for census datas, in particular the relsurv::slopop dataset for slovenia :

v1 = daily_hazard(slopop, a, d, s)
v2 = daily_hazard(slopop, a, d; sex=s)
v3 = daily_hazard(slopop[s], a, d)
(v1,v2,v3)

(2.20613869487702e-6, 2.20613869487702e-6, 2.20613869487702e-6)

Sometimes there is more covariates in the datasets than only the sex. Ths is for example the case for the survival::survexp.usr dataset that includes the race. In this case, the calling structure is very similar:

r = :white
v1 = daily_hazard(survexp_usr, a, d, s, r)
v2 = daily_hazard(survexp_usr, a, d; sex=s, race=r)
v3 = daily_hazard(survexp_usr[s, r], a, d)
(v1,v2,v3)

(2.8872165601933327e-6, 2.8872165601933327e-6, 2.8872165601933327e-6)

Fetching these daily hazard is a very sensitive operation that should be as fast as possible since it is usually used in the middle of very-hot loops. Therefore, we take care of the performance of our fetching algorithms.

Check out the folliwng index for a list of availiable ratetables:

RateTables.hmd_countries

RateTables.hmd_rates

RateTables.slopop

RateTables.survexp_fr

RateTables.survexp_mn

RateTables.survexp_us

RateTables.survexp_usr

RateTables.RateTable

RateTables.daily_hazard

RateTables.hmd_countries — Constant

`hmd_countries`

Gives details about the countries codes used in the hmd_rates dataset.

RateTables.hmd_rates — Constant

`hmd_rates`

RateTable providing daily hazard rates for both sexes for several countries. They are derived from annual death probabilities (qₓ's) from the Human Mortality Database at https://www.mortality.org/

Segmented by country ∈ ... and sex ∈ (:male, :female, :total).

Dayly hazard for a slovenian male, born on the 1rst of january 1927, 192 days after his 37th birthday can for example be obtained as follows:

daily_hazard(hmd_rates[:svn, :male], 37*RT_DAYS_IN_YEAR + 192, 1927*RT_DAYS_IN_YEAR + 192, 1)

The list of countries codes is given with details in the hmd_countries constant.

RateTables.slopop — Constant

`slopop`

Slovene census data. Correspond to R's package relsurv slopop ratetable. Segmented by sex ∈ (:male, :female).

RateTables.survexp_fr — Constant

`survexp_fr`

French census datas, drawn from R's package survexp.fr. Death rates are available from 1977 to 2019 for males and females aged from 0 to 99. Segmented by sex ∈ (:male, :female)

Source: https://www.insee.fr/fr/statistiques/fichier/5390366/fm_t68.xlsx

References: Institut National de la Statistique et des Etudes Economiques

RateTables.survexp_mn — Constant

`survexp_mn`

Census data set for the US population, drawn from R's package survival. RateTable survexp_mn gives total Minnesota population, by age and sex, 1970 to 2013. Segmented by sex ∈ (:male, :female)

RateTables.survexp_us — Constant

`survexp_us`

Census data set for the US population, drawn from R's package survival. RateTable survexp_us gives total United States population, by age and sex, 1940 to 2012. Segmented by sex ∈ (:male, :female)

RateTables.survexp_usr — Constant

`survexp_usr`

Census data set for the US population, drawn from R's package survival. RateTable survexp_usr gives the United States population, by age, sex and race, 1940 to 2014. Race is white or black. For 1960 and 1970 the black population values were not reported separately, so the nonwhite values were used. (Over the years, the reported tables have differed wrt reporting non-white and/or black.). Segmented by sex ∈ (:male, :female) and race∈ (:white, :black)`.

RateTables.RateTable — Type

`RateTable`

This class contains daily rate tables used in person-years computation.

Each of these tables contains the daily hazard rate for a matched subject from the population, defined as $-\log(1-qₓ)$ for $qₓ$ the 1 year probability of death as reported in the original tables from the US Census. The tables are given in terms of hazard per day for computational convenience.

RateTables.daily_hazard — Method

`daily_hazard(rt::BasicRateTable,age,date)`

This function queries daily hazard values from a given BasicRateTable. The parameters age and date have to be in days (1 year = 365.241 days).