RateTables

Documentation for RateTables. This package provides rate tables for person-year computations, alike R's ratetable class.

You can install it through :

using Pkg
Pkg.add("https://github.com/JuliaSurv/RateTables.jl")

Then loading this package provides a few constants rate tables objects. See the bottom of this page for the available rate tables. For the sake of the example, we'll use the RT_HMD dictionary, which stores contries rates tables extracted from the Human Mortality Database (HMD):

using RateTables
rt = RT_HMD[:SVN] # SVN for slovenia.

RT(:sex,) with elements BRT(0..110 × 1983..2019).

To obtain, for example, the daily hazard rate for a male, slovene, on its 20th birthday on the first of january 2010, you can call the daily_hazard function. It needs to be called with the arguments in a specific format:

The age needs to be given in days, and the converting factor is 1 year = 365.241 days.

The date also needs to be given in days, same converting factor.

Other covariates formats depend on the rate table. Usually, there is a sex covariates with values :male and :female, sometimes :total too, but other covariates might have different format depending on the rate table.

Then the right calling syntax is as follows:

a = 20*365.241
d = 2010*365.241
s = :male
v1 = daily_hazard(rt, a, d, s)
v2 = daily_hazard(rt, a, d; sex=s)
v3 = daily_hazard(rt[s], a, d)
(v1,v2,v3)

(2.092569007129928e-6, 2.092569007129928e-6, 2.092569007129928e-6)

Note that there is a discrepency between HMD datasets and datasets from other places. For completeness, this package also include datasets commonly used in R for census datas, in particular the relsurv::slopop dataset for slovenia :

v1 = daily_hazard(slopop, a, d, s)
v2 = daily_hazard(slopop, a, d; sex=s)
v3 = daily_hazard(slopop[s], a, d)
(v1,v2,v3)

(2.20613869487702e-6, 2.20613869487702e-6, 2.20613869487702e-6)

Sometimes there is more covariates in the datasets than only the sex. Ths is for example the case for the survival::survexp.usr dataset that includes the race. In this case, the calling structure is very similar:

r = :white
v1 = daily_hazard(survexp_usr, a, d, s, r)
v2 = daily_hazard(survexp_usr, a, d; sex=s, race=r)
v3 = daily_hazard(survexp_usr[s, r], a, d)
(v1,v2,v3)

(2.8872165601933327e-6, 2.8872165601933327e-6, 2.8872165601933327e-6)

Fetching these daily hazard is a very sensitive operation that should be as fast as possible since it is usually used in the middle of very-hot loops. Therefore, we take care of the performance of our fetching algorithms.

Check out the index below for the list of availiable datasets.

RateTables.RT_HMD

RateTables.slopop

RateTables.survexp_fr

RateTables.survexp_mn

RateTables.survexp_us

RateTables.survexp_usr

RateTables.RateTable

RateTables.daily_hazard

RateTables.RT_HMD — Constant

`RT_HMD`

Dictionary that contains the rates tables for each country. The country codes are used to differentiate them.

These rates tables contain daily hazard rates for both sexes. They are derived from annual death probabilities (qₓ's) from the Human Mortality Database at https://www.mortality.org/

Dayly hazard for a slovenian male (code 0), born on the 1rst of january 1927, 192 days after his 37th birthday can for example be obtained as follows:

solovia_rt = RT_HMD[:SVN] daily_hazard(slovenia_rt, 37*RT_DAYS_IN_YEAR + 192, 1927*RT_DAYS_IN_YEAR + 192, 1)

The list of availiables countries can be queried via:

Dict(k => RT_HMD[k].country_name for k in keys(RT_HMD))

RateTables.slopop — Constant

`slopop`

This ratetable corresponds to R's package relsurv, drawn from census data set for the Slovene population.

RateTables.survexp_fr — Constant

`survexp_fr`

French census datas, drawn from R's package survexp.fr. Death rates are available from 1977 to 2019 for males and females aged from 0 to 99

Source: https://www.insee.fr/fr/statistiques/fichier/5390366/fm_t68.xlsx

References: Institut National de la Statistique et des Etudes Economiques

RateTables.survexp_mn — Constant

`survexp_us`
`survexp_usr`
`survexp_mn`

Census data sets for the US population, drawn from R's package survival.

survexp_us: total United States population, by age and sex, 1940 to 2012.

survexp_usr: United States population, by age, sex and race, 1940 to 2014. Race is white or black. For 1960 and 1970 the black population values were not reported separately, so the nonwhite values were used. (Over the years, the reported tables have differed wrt reporting non-white and/or black.)

survexp_mn: total Minnesota population, by age and sex, 1970 to 2013.

Each of these tables contains the daily hazard rate for a matched subject from the population, defined as $-\log(1-qₓ)$ for $qₓ$ the 1 year probability of death as reported in the original tables from the US Census. For age 25 in 1970, for instance, $pₓ=1−qₓ$ is is the probability that a subject who becomes 25 years of age in 1970 will achieve his/her 26th birthday. The tables are recast in terms of hazard per day for computational convenience.

age and dates dimensions are in days.

RateTables.survexp_us — Constant

`survexp_us`
`survexp_usr`
`survexp_mn`

Census data sets for the US population, drawn from R's package survival.

survexp_us: total United States population, by age and sex, 1940 to 2012.

survexp_usr: United States population, by age, sex and race, 1940 to 2014. Race is white or black. For 1960 and 1970 the black population values were not reported separately, so the nonwhite values were used. (Over the years, the reported tables have differed wrt reporting non-white and/or black.)

survexp_mn: total Minnesota population, by age and sex, 1970 to 2013.

Each of these tables contains the daily hazard rate for a matched subject from the population, defined as $-\log(1-qₓ)$ for $qₓ$ the 1 year probability of death as reported in the original tables from the US Census. For age 25 in 1970, for instance, $pₓ=1−qₓ$ is is the probability that a subject who becomes 25 years of age in 1970 will achieve his/her 26th birthday. The tables are recast in terms of hazard per day for computational convenience.

age and dates dimensions are in days.

RateTables.survexp_usr — Constant

`survexp_us`
`survexp_usr`
`survexp_mn`

Census data sets for the US population, drawn from R's package survival.

survexp_us: total United States population, by age and sex, 1940 to 2012.

survexp_usr: United States population, by age, sex and race, 1940 to 2014. Race is white or black. For 1960 and 1970 the black population values were not reported separately, so the nonwhite values were used. (Over the years, the reported tables have differed wrt reporting non-white and/or black.)

survexp_mn: total Minnesota population, by age and sex, 1970 to 2013.

Each of these tables contains the daily hazard rate for a matched subject from the population, defined as $-\log(1-qₓ)$ for $qₓ$ the 1 year probability of death as reported in the original tables from the US Census. For age 25 in 1970, for instance, $pₓ=1−qₓ$ is is the probability that a subject who becomes 25 years of age in 1970 will achieve his/her 26th birthday. The tables are recast in terms of hazard per day for computational convenience.

age and dates dimensions are in days.

RateTables.RateTable — Type

RateTable

This class deals with daily rate tables used in person-years computation. For the moment it simply holds data from the HMD.

RateTables.daily_hazard — Method

`daily_hazard(rt::BasicRateTable,age,date)`

This function queries daily hazard values from a given BasicRateTable. The parameters age and date have to be in days (1 year = 365.241 days).